+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Added
 - Arquitens-class command cruiser to Galactic Empire
 - Arquitens-class light cruiser, Haven-class medical station and IPV-2C Stealth Corvette to Galactic Republic
+- New faction New Republic
+- Starhawk-class Battleship to New Republic
 
 ### Changed